Implement simple in-memory TTL LRU cache for LDAP user groups response.
This decreased the load on LDAP server because it means that when user
authenticated with StackStorm only a single request to LDAP server will be
issued (authenticate + get groups), previously there it results in two
connections being established and two requests being issued to the LDAP
server.

In addition to that, it also reduces load on LDAP server when the same
user tries to concurrently authenticate with StackStorm because it will
only result in N authenticate requests to LDAP server while previously
it results in N * 2 requests to LDAP server (1 for authenticate, 1 for
get user groups).