loosen version requirements (#108)
I could not build python-ldap on one machine that had openldap 2.6.
And the pinned version of cachetools is from 2017 and does not
technically support python3.8 until cachetools v3.1.

So, loosen the requirements. Let the lockfile in the st2 repo
handle pinning a precise version. This is a libraray, so it should
not have deps locked down this far anyway. Deps should be locked
at an app-level, not at a library level.